Move clipboard class into own file
