+++ Clipboard.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <string>
+
+namespace Linter
+{
+    class Clipboard
+    {
+      public:
+        explicit Clipboard(HWND self);
+
+        Clipboard(Clipboard const &) = delete;
+        Clipboard(Clipboard &&) = delete;
+
+        Clipboard operator=(Clipboard const &) = delete;
+        Clipboard operator=(Clipboard &&) = delete;
+
+        ~Clipboard();
+
+        void empty();
+
+        void copy(std::wstring const &str);
+
+      private:
+        HGLOBAL mem_handle_ = nullptr;
+    };
+
+}    // namespace Linter
